Move get_delimiter_from_filepath to utils
