+++ src/pkg/caendr/caendr/models/job_pipeline/heritability_pipeline.py
@@ -9,8 +9,9 @@ from caendr.models.run             import HeritabilityRunner
 # Services
 from caendr.models.datastore       import Species
 from caendr.models.error           import DataFormatError
-from caendr.services.validate      import get_delimiter_from_filepath, validate_file, NumberValidator, StrainValidator, TraitValidator
+from caendr.services.validate      import validate_file, NumberValidator, StrainValidator, TraitValidator
 from caendr.services.cloud.storage import upload_blob_from_file
+from caendr.utils.data             import get_delimiter_from_filepath
 from caendr.utils.env              import get_env_var
 from caendr.utils.file             import get_file_hash
